Sign in bug fixed

+++ src/controller/Auth.js
@@ -40,32 +40,29 @@ export async function signIn(req, res) {
     const checkUser = await db.collection("users").findOne({ email });
 
     if (!checkUser) return res.status(400).send("Usuário ou senha incorretos");
-
     const isCorrectPassword = bcrypt.compareSync(password, checkUser.password);
 
     if (!isCorrectPassword) {
       return res.status(400).send("Usuário ou senha incorretos");
     }
     if (token) {
-      const removableSession = await db.collection("sessions").findOne({ idUser: checkUser._id })
-      if (removableSession) {
+      const removableSession = await db.collection("sessions").findOne({ idUser: ObjectId(checkUser._id )})
+      if (removableSession && (removableSession.token !== token)) {
         await db.collection("sessions").deleteOne({ idUser: checkUser._id });
       }
-
-      const session = await db.collection("sessions").findOne({ token });
-      await db.collection("sessions").updateOne({ _id: ObjectId(session._id) }, { $set: { idUser: checkUser._id } });
-
+      const session = await db.collection("sessions").findOne({ token:token });
+      const newSession = await db.collection("sessions").updateOne({ token }, { $set: { idUser: checkUser._id } });
       const previousCart = await db.collection("carts").findOne({ idUser: checkUser._id });
       let previousCartItems = [];
       if (previousCart) {
-        const previousCartItems = previousCart.chosenItems;
+        previousCartItems = previousCart.chosenItems;
         await db.collection("carts").deleteOne({ idUser: checkUser._id });
       }
       const currentCart = await db.collection("carts").findOne({ token });
       const currentCartItems = currentCart.chosenItems;
       const items = [...currentCartItems, ...previousCartItems];
-      await db.collection("carts").updateOne({ token }, { $set: { chosenItems: [...items] } })
-      await db.collection("carts").updateOne({ idUser }, { $set: { chosenItems: [...items] } })
+      await db.collection("carts").deleteOne({ idUser: checkUser._id });
+      await db.collection("carts").updateOne({ token }, { $set: { chosenItems: [...items], idUser:checkUser._id  } })
       return res.status(200).send("Ok");
     }
   } catch (error) {
